created Stopwatch.js and moved stopwatch comp
